% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/korhesap.R
\name{korhesap}
\alias{korhesap}
\title{Korelason Hesaplıyoruz
tidy fonksiyonu ile korelasyon hesaplıyoruz
bu paket çok basit iki sürekli değişken arasındaki ilişkiyi belirlemeye yarar.}
\usage{
korhesap(data, var1, var2)
}
\arguments{
\item{data}{yüklemeniz gereken veri seti}

\item{var1}{birinci değişken}

\item{var2}{ikinci değişken}
}
\value{
Bir tibble tablosu ile korelasyon sonuçlarını verir
}
\description{
Korelason Hesaplıyoruz
tidy fonksiyonu ile korelasyon hesaplıyoruz
bu paket çok basit iki sürekli değişken arasındaki ilişkiyi belirlemeye yarar.
}
\examples{
korhesap(data = faithful, var1 = eruptions, var2 = waiting)
}
